+++ src/cwe_checker_lib/src/intermediate_representation/term/builder.rs
@@ -7,6 +7,38 @@ use crate::{expr, intermediate_representation::*, variable};
 #[cfg(test)]
 use std::collections::{BTreeMap, BTreeSet};
 
+#[cfg(test)]
+impl Expression {
+    /// Shortcut for creating a cast expression
+    #[cfg(test)]
+    pub fn cast(self, op: CastOpType) -> Expression {
+        Expression::Cast {
+            op,
+            size: ByteSize::new(8),
+            arg: Box::new(self),
+        }
+    }
+
+    /// Shortcut for creating a subpiece expression
+    #[cfg(test)]
+    pub fn subpiece(self, low_byte: ByteSize, size: ByteSize) -> Expression {
+        Expression::Subpiece {
+            low_byte,
+            size,
+            arg: Box::new(self),
+        }
+    }
+
+    /// Shortcut for creating unary operation expressions.
+    #[cfg(test)]
+    pub fn un_op(self, op: UnOpType) -> Expression {
+        Expression::UnOp {
+            op,
+            arg: Box::new(self),
+        }
+    }
+}
+
 /// ## Helper functions for building defs
 #[cfg(test)]
 impl Def {
